changed frontend for better role detection and server-client components
