+++ app/(protected)/presets/presets-page-client.tsx
@@ -1,0 +1,57 @@
+"use client";
+
+import PresetColumns from "@/components/preset-columns";
+import { useAuthData } from "@/lib/context/auth-context";
+
+interface PresetsPageClientProps {
+  presetsData: {
+    apparatusPresets: any[];
+    sessionPresets: any[];
+    microcyclePresets: any[];
+    microcyclesSessionsPresets: any[];
+    macrocyclePresets: any[];
+    macrocyclesMicrocyclesPresets: any[];
+  };
+}
+
+export default function PresetsPageClient({
+  presetsData,
+}: PresetsPageClientProps) {
+  const { role, loading } = useAuthData();
+
+  if (loading) {
+    return (
+      <div className="flex h-96 items-center justify-center">
+        <p>Caricamento...</p>
+      </div>
+    );
+  }
+
+  if (role !== "coach") {
+    return <div>Solo i tecnici possono gestire i preset.</div>;
+  }
+
+  const {
+    apparatusPresets,
+    sessionPresets,
+    microcyclePresets,
+    microcyclesSessionsPresets,
+    macrocyclePresets,
+    macrocyclesMicrocyclesPresets,
+  } = presetsData;
+
+  return (
+    <div className="container mx-auto p-6">
+      <h1 className="mb-6 text-2xl font-bold">Gestione Preset</h1>
+
+      <PresetColumns
+        apparatusPresets={apparatusPresets}
+        sessionPresets={sessionPresets}
+        microcyclePresets={microcyclePresets}
+        microcyclesSessionsPresets={microcyclesSessionsPresets}
+        macrocyclePresets={macrocyclePresets}
+        macrocyclesMicrocyclesPresets={macrocyclesMicrocyclesPresets}
+      />
+    </div>
+  );
+}
